requirements.txt
gradio>=4.0.0
mysql-connector-python>=8.0.0